feat: delta-neutral funding-carry research lane (Sharpe 8.15 dev / 6.88 full)

+++ src/quant_fund/backtest/perp_engine.py
@@ -83,7 +83,7 @@ class PerpBook:
         return sum(self.qty[s] * marks[s] for s in self.qty if s in marks)
 
 
-def infer_periods_per_year(times: list[datetime], hint_seconds: int | None = None) -> float:
+def infer_periods_per_year(times: list[datetime], hint_seconds: float | None = None) -> float:
     """Bar-spacing annualization on a 365.25-day calendar.
 
     Median spacing is robust to isolated missing bars; a hint wins only when
@@ -175,9 +175,7 @@ def run_perp_backtest(
     # Pre-group target rows by event_time once — a per-bar frame filter is
     # O(N) per timestamp and dominates runtime on hourly panels.
     wmap_by_time: dict[datetime, dict[str, float]] = {}
-    for wrow in weights.select("event_time", "security_id", "target_weight").iter_rows(
-        named=True
-    ):
+    for wrow in weights.select("event_time", "security_id", "target_weight").iter_rows(named=True):
         wmap_by_time.setdefault(wrow["event_time"], {})[str(wrow["security_id"])] = float(
             wrow["target_weight"]
         )
